refactor: 이것저것 리팩토링 겸 삭제 겸 뭐 등등
