+++ src/main/java/com/example/hospitalreservation/common/exception/ExceptionCode.java
@@ -4,8 +4,6 @@ import org.springframework.http.HttpStatus;
 
 public interface ExceptionCode {
 
-    String getName();
-
     HttpStatus getHttpStatus();
 
     String getCode();
