+++ src/main/java/com/example/hospitalreservation/patient/domain/exception/PatientExceptionCode.java
@@ -19,11 +19,6 @@ public enum PatientExceptionCode implements ExceptionCode {
     }
 
     @Override
-    public String getName() {
-        return "";
-    }
-
-    @Override
     public HttpStatus getHttpStatus() {
         return httpStatus;
     }
