--- src/main/resources/templates/nyah.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>메롱 페이지</title>
-    <link rel="stylesheet" type="text/css" href="/css/style.css">
-</head>
-<body>
-<h1>메롱이다 ~</h1>
-<br>
-<a href="/reservations">예약 목록으로 돌아가기</a>
-</body>
-</html>
